Split out log server into separate module
asyncio is not available on Python < 3.5.
To make cdist still usable on these older versions the log server is split out
to a separate module that is only loaded when required.
